Move rolebased test to common place, update changelog
